requirements.txt
requests~=2.23.0
python-telegram-bot==13.3
numpy~=1.19.5